Doors troubleshoot (#389)
* add two extensions for doors, DoorVariant::IsBreakable() && DoorVariant::GetNametag()

* Fixed door type stuff (#390)

* remove the setter of the Sever::Port property

It's impossible to switch to another port without restarting the server or a port listener (LiteNetLib4Mirror is this case) itself.
Anyway, if a user wants to make that kind of shit, let they do, but not provide it as part of the API.

* Fixed door type stuff


* merge the DoorTypeExtension class into the DoorExtensions class

* remove duplicated 'for' from the summary of the RegisterDoorTypesOnLevelLoad method
