+++ Exiled.API/Enums/DoorType.cs
@@ -7,11 +7,20 @@
 
 namespace Exiled.API.Enums
 {
+    using System;
+
+#pragma warning disable SA1124 // Do not use regions
+
     /// <summary>
     /// Unique identifier for the different types of doors.
     /// </summary>
     public enum DoorType
     {
+        /// <summary>
+        /// Represents an unknown door.
+        /// </summary>
+        UnknownDoor = 0,
+
         /// <summary>
         /// Represents the 012 door.
         /// </summary>
@@ -63,9 +72,14 @@ namespace Exiled.API.Enums
         Scp106Secondary,
 
         /// <summary>
-        /// Represents the 173 door.
+        /// Represents the 173 gate.
         /// </summary>
-        Scp173,
+        Scp173Gate,
+
+        /// <summary>
+        /// Represents the door between the 173 gate and the 173 armory.
+        /// </summary>
+        Scp173Connector,
 
         /// <summary>
         /// Represents the 173_ARMORY door.
@@ -78,19 +92,14 @@ namespace Exiled.API.Enums
         Scp173Bottom,
 
         /// <summary>
-        /// Represents the 372 door.
+        /// Represents the GR18 door.
         /// </summary>
-        Scp372,
+        GR18,
 
         /// <summary>
         /// Represents the 914 door.
         /// </summary>
         Scp914,
-
-        /// <summary>
-        /// Represents the Airlocks door.
-        /// </summary>
-        Airlocks,
 
         /// <summary>
         /// Represents the CHECKPOINT_ENT door.
@@ -108,24 +117,24 @@ namespace Exiled.API.Enums
         CheckpointLczB,
 
         /// <summary>
-        /// Represents the ContDoor door.
-        /// </summary>
-        ContDoor,
-
-        /// <summary>
-        /// Represents the EntrDoor door.
+        /// RRepresents any entrance zone styled door.
         /// </summary>
         EntranceDoor,
 
         /// <summary>
-        /// Represents the ESCAPE door.
+        /// Represents the ESCAPE_PRIMARY door.
         /// </summary>
-        Escape,
+        EscapePrimary,
 
         /// <summary>
-        /// Represents the ESCAPE_INNER door.
+        /// Represents the ESCAPE_SECONDARY door.
         /// </summary>
-        EscapeInner,
+        EscapeSecondary,
+
+        /// <summary>
+        /// Represents the SERVERS_BOTTOM door.
+        /// </summary>
+        ServersBottom,
 
         /// <summary>
         /// Represents the GATE_A door.
@@ -143,7 +152,7 @@ namespace Exiled.API.Enums
         HczArmory,
 
         /// <summary>
-        /// Represents the HeavyContainmentDoor door.
+        /// Represents any heavy containment styled door.
         /// </summary>
         HeavyContainmentDoor,
 
@@ -183,7 +192,7 @@ namespace Exiled.API.Enums
         LczWc,
 
         /// <summary>
-        /// Represents the LightContainmentDoor door.
+        /// Represents any light containment styled door.
         /// </summary>
         LightContainmentDoor,
 
@@ -207,9 +216,44 @@ namespace Exiled.API.Enums
         /// </summary>
         SurfaceGate,
 
+        #region Deprecated
+
         /// <summary>
-        /// Represents an unknown door.
+        /// Represents the 372 door.
         /// </summary>
-        UnknownDoor,
+        [Obsolete("Replaced with GR18.")]
+        Scp372 = GR18,
+
+        /// <summary>
+        /// Represents the Airlocks door.
+        /// </summary>
+        [Obsolete("Unused.")]
+        Airlocks = -1,
+
+        /// <summary>
+        /// Represents the ContDoor door.
+        /// </summary>
+        [Obsolete("Unused.")]
+        ContDoor = -1,
+
+        /// <summary>
+        /// Represents the ESCAPE door.
+        /// </summary>
+        [Obsolete("Replaced with EscapePrimary.")]
+        Escape = EscapePrimary,
+
+        /// <summary>
+        /// Represents the ESCAPE_INNER door.
+        /// </summary>
+        [Obsolete("Replaced with EscapeSecondary.")]
+        EscapeInner = EscapeSecondary,
+
+        /// <summary>
+        /// Represents the 173 door.
+        /// </summary>
+        [Obsolete("Replaced with Scp173Conenctor")]
+        Scp173 = Scp173Connector,
+
+        #endregion
     }
 }
